Enable running the Project with http host without redirection
Previously, the project only worked with the default HTTP host, and
would redirect to it if a custom host was used. With this change, the
project can now run with a custom HTTP host without redirection. This
was achieved by updating the project configuration to allow for the
custom host. This will improve the flexibility of the project and make
it easier to use in various environments.